@@ -23,7 +23,6 @@
   ],
   "devDependencies": {
     "@types/jest": "^26.0.15",
-    "@types/react-dropzone": "^3.13.4",
     "@typescript-eslint/eslint-plugin": "^5.14.0",
     "@typescript-eslint/eslint-plugin-tslint": "^5.14.0",
     "@typescript-eslint/parser": "^5.14.0",
@@ -81,9 +80,10 @@
     "@types/markdown-it": "^0.0.4",
     "@types/moment-duration-format": "^2.2.2",
     "@types/query-string": "^5.0.1",
-    "@types/react": "^17.0.39",
+    "@types/react": "^18.0.5",
     "@types/react-autosuggest": "^10.1.5",
-    "@types/react-dom": "^17.0.13",
+    "@types/react-dom": "^18.0.1",
+    "@types/react-dropzone": "^3.13.4",
     "@types/react-router-dom": "^5.3.2",
     "@types/react-select": "^3.0.26",
     "@types/react-virtualized": "^9.18.8",
@@ -101,11 +101,11 @@
     "moment-timezone": "^0.5.31",
     "prop-types": "^15.5.0",
     "query-string": "^5.0.1",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-autosuggest": "^10.1.0",
     "react-charts": "^2.0.0-beta.6",
     "react-datetime": "^3.0.4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-dropzone": "^3.7.3",
     "react-linkify": "^1.0.0-alpha",
     "react-number-format": "^3.0.2",
